Split app/util/hooks.tsx into separate hook files
Separate hook files are easier to isolate and maintain
