+++ app/components/MessagePathSyntax/useCachedGetMessagePathDataItems.ts
@@ -17,11 +17,13 @@ import { useCallback, useMemo, useRef } from "react";
 import * as PanelAPI from "@foxglove-studio/app/PanelAPI";
 import { isTypicalFilterName } from "@foxglove-studio/app/components/MessagePathSyntax/isTypicalFilterName";
 import parseRosPath from "@foxglove-studio/app/components/MessagePathSyntax/parseRosPath";
+import useChangeDetector from "@foxglove-studio/app/hooks/useChangeDetector";
+import useDeepMemo from "@foxglove-studio/app/hooks/useDeepMemo";
 import useGlobalVariables, { GlobalVariables } from "@foxglove-studio/app/hooks/useGlobalVariables";
+import useShallowMemo from "@foxglove-studio/app/hooks/useShallowMemo";
 import { Message, ReflectiveMessage, Topic } from "@foxglove-studio/app/players/types";
 import { RosDatatypes } from "@foxglove-studio/app/types/RosDatatypes";
 import { fieldNames, getField, getIndex } from "@foxglove-studio/app/util/binaryObjects";
-import { useChangeDetector, useDeepMemo, useShallowMemo } from "@foxglove-studio/app/util/hooks";
 import {
   enumValuesByDatatypeAndField,
   getTopicsByTopicName,
