+++ app/components/MessagePathSyntax/useMessagesByPath.ts
@@ -1,6 +1,16 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
+//
+// This file incorporates work covered by the following copyright and
+// permission notice:
+//
+//   Copyright 2018-2021 Cruise LLC
+//
+//   This source code is licensed under the Apache License, Version 2.0,
+//   found at http://www.apache.org/licenses/LICENSE-2.0
+//   You may not use this file except in compliance with the License.
+
 import { useMemo } from "react";
 
 import * as PanelAPI from "@foxglove-studio/app/PanelAPI";
@@ -9,7 +19,7 @@ import {
   MessageDataItemsByPath,
   useDecodeMessagePathsForMessagesByTopic,
 } from "@foxglove-studio/app/components/MessagePathSyntax/useCachedGetMessagePathDataItems";
-import { useShallowMemo } from "@foxglove-studio/app/util/hooks";
+import useShallowMemo from "@foxglove-studio/app/hooks/useShallowMemo";
 
 // Given a set of message paths, subscribe to the appropriate topics and return
 // messages with their queried data decoded for each path.
