+++ app/hooks/useShallowMemo.ts
@@ -1,0 +1,16 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+import { useRef } from "react";
+import shallowequal from "shallowequal";
+
+// Continues to return the same instance as long as shallow equality is maintained.
+
+export default function useShallowMemo<T>(value: T): T {
+  const ref = useRef(value);
+  if (shallowequal(value, ref.current)) {
+    return ref.current;
+  }
+  ref.current = value;
+  return value;
+}
