+++ app/panels/Plot/index.tsx
@@ -34,6 +34,7 @@ import {
   getTooltipItemForMessageHistoryItem,
   TooltipItem,
 } from "@foxglove-studio/app/components/TimeBasedChart";
+import useShallowMemo from "@foxglove-studio/app/hooks/useShallowMemo";
 import PlotChart, {
   getDatasetsAndTooltips,
   PlotDataByPath,
@@ -41,7 +42,6 @@ import PlotChart, {
 import PlotLegend from "@foxglove-studio/app/panels/Plot/PlotLegend";
 import PlotMenu from "@foxglove-studio/app/panels/Plot/PlotMenu";
 import { PanelConfig } from "@foxglove-studio/app/types/panels";
-import { useShallowMemo } from "@foxglove-studio/app/util/hooks";
 import { fromSec, subtractTimes, toSec } from "@foxglove-studio/app/util/time";
 
 import helpContent from "./index.help.md";
