+++ app/panels/ThreeDimensionalViz/TopicTree/useTopicTree.ts
@@ -15,6 +15,7 @@ import { difference, keyBy, uniq, mapValues, xor, isEqual, flatten, omit } from 
 import { useMemo, useCallback, useRef, createContext } from "react";
 import { useDebounce } from "use-debounce";
 
+import useShallowMemo from "@foxglove-studio/app/hooks/useShallowMemo";
 import { TOPIC_DISPLAY_MODES } from "@foxglove-studio/app/panels/ThreeDimensionalViz/TopicTree/TopicViewModeSelector";
 import filterMap from "@foxglove-studio/app/util/filterMap";
 import {
@@ -22,7 +23,6 @@ import {
   FOXGLOVE_GRID_TOPIC,
   SECOND_SOURCE_PREFIX,
 } from "@foxglove-studio/app/util/globalConstants";
-import { useShallowMemo } from "@foxglove-studio/app/util/hooks";
 
 import {
   TreeNode,
